[DERCBOT-542]  - découpage des verticles
