+++ bot/admin/server/src/main/kotlin/verticle/StoryVerticle.kt
@@ -16,8 +16,14 @@
 
 package ai.tock.bot.admin.verticle
 
+import ai.tock.bot.admin.model.BotStoryDefinitionConfiguration
+import ai.tock.bot.admin.model.CreateStoryRequest
+import ai.tock.bot.admin.model.StorySearchRequest
+import ai.tock.bot.admin.service.BotAdminService
 import ai.tock.bot.admin.service.StoryService
+import ai.tock.bot.admin.story.dump.StoryDefinitionConfigurationDump
 import ai.tock.bot.bean.TickStoryQuery
+import ai.tock.nlp.front.client.FrontClient
 import ai.tock.shared.exception.admin.AdminException
 import ai.tock.shared.security.TockUser
 import ai.tock.shared.security.TockUserRole
@@ -43,10 +49,8 @@ class StoryVerticle: ChildVerticle<AdminException> {
     /**
      * Declaration of routes and association to the appropriate handler
      */
-    override fun configure(verticle: WebVerticle<AdminException>) {
-        logger.info { "configure StoryVerticle" }
-
-        with(verticle) {
+    override fun configure(parent: WebVerticle<AdminException>) {
+        with(parent) {
 
             val createTickStory: (RoutingContext, TickStoryQuery) -> Unit = { context, tickStory ->
                 Companion.logger.debug { "request to create tick story <${tickStory.storyId}>" }
@@ -68,7 +72,116 @@ class StoryVerticle: ChildVerticle<AdminException> {
                 setOf(TockUserRole.botUser, TockUserRole.faqBotUser),
                 handler = toRequestHandler(deleteStoryByStoryDefinitionConfigurationId)
             )
+
+            blockingJsonGet("/bot/story/:appName/export", TockUserRole.botUser, handler= toRequestHandler { context ->
+                BotAdminService.exportStories(context.organization, context.path("appName"))
+            })
+
+            blockingJsonGet("/bot/story/:appName/export/:storyConfigurationId",
+                TockUserRole.botUser, handler= toRequestHandler { context ->
+                val exportStory = BotAdminService.exportStory(
+                    context.organization,
+                    context.path("appName"),
+                    context.path("storyConfigurationId")
+                )
+                exportStory?.let { listOf(it) } ?: emptyList()
+            })
+
+            blockingJsonPost(
+                "/bot/story/load",
+                setOf(TockUserRole.botUser, TockUserRole.faqBotUser),
+                handler = toRequestHandler { context, request: StorySearchRequest ->
+                    if (context.organization == request.namespace) {
+                        BotAdminService.loadStories(request)
+                    } else {
+                        WebVerticle.unauthorized()
+                    }
+                })
+
+            blockingJsonPost("/bot/story/search",
+                setOf(TockUserRole.botUser, TockUserRole.faqBotUser),
+                handler = toRequestHandler { context, request: StorySearchRequest ->
+                    if (context.organization == request.namespace) {
+                        BotAdminService.searchStories(request)
+                    } else {
+                        WebVerticle.unauthorized()
+                    }
+                })
+
+            blockingJsonGet("/bot/story/:storyId",
+                setOf(TockUserRole.botUser, TockUserRole.faqBotUser),
+                handler = toRequestHandler { context ->
+                    BotAdminService.findStory(context.organization, context.path("storyId"))
+                })
+
+            blockingJsonGet("/bot/story/:botId/settings",
+                TockUserRole.botUser,
+                handler = toRequestHandler { context ->
+                    BotAdminService.findRuntimeStorySettings(context.organization, context.path("botId"))
+                })
+
+            blockingJsonGet("/bot/story/:botId/:intent",
+                TockUserRole.botUser,
+                handler = toRequestHandler { context ->
+                    BotAdminService.findConfiguredStoryByBotIdAndIntent(
+                        context.organization,
+                        context.path("botId"),
+                        context.path("intent")
+                    )
+                })
+
+            blockingJsonPost(
+                "/bot/story/new",
+                setOf(TockUserRole.botUser, TockUserRole.faqBotUser),
+                requestLogger<CreateStoryRequest, AdminException>("Create Story") { context, r ->
+                    r?.story?.let { s ->
+                        BotAdminService.getBotConfigurationsByNamespaceAndBotId(context.organization, s.botId)
+                            .firstOrNull()
+                            ?.let {
+                                FrontClient.getApplicationByNamespaceAndName(
+                                    context.organization,
+                                    it.nlpModel
+                                )?._id
+                            }
+                    }
+                },
+                handler = toRequestHandler { context, query: CreateStoryRequest ->
+                    BotAdminService.createStory(context.organization, query, context.userLogin) ?: WebVerticle.unauthorized()
+                })
+
+            blockingUploadJsonPost(
+                "/bot/story/:appName/:locale/import",
+                TockUserRole.botUser,
+                simpleLogger("JSON Import Response Labels"),
+                handler = toRequestHandler { context, stories: List<StoryDefinitionConfigurationDump> ->
+                    BotAdminService.importStories(
+                        context.organization,
+                        context.path("appName"),
+                        context.pathToLocale("locale"),
+                        stories,
+                        context.userLogin
+                    )
+                })
+
+            blockingJsonPost(
+                "/bot/story",
+                setOf(TockUserRole.botUser, TockUserRole.faqBotUser),
+                requestLogger<BotStoryDefinitionConfiguration, AdminException>("Update Story") { context, r ->
+                    r?.let { s ->
+                        BotAdminService.getBotConfigurationsByNamespaceAndBotId(context.organization, s.botId)
+                            .firstOrNull()
+                            ?.let {
+                                FrontClient.getApplicationByNamespaceAndName(
+                                    context.organization,
+                                    it.nlpModel
+                                )?._id
+                            }
+                    }
+                },
+                handler = toRequestHandler { context, story: BotStoryDefinitionConfiguration ->
+                    BotAdminService.saveStory(context.organization, story, context.userLogin) ?: WebVerticle.unauthorized()
+                })
         }
-            // TODO : migrate all story vert.x handlers to here  (ex BotAdminVerticle)
+
     }
 }
